refactor(useGrantedPolicies): use provider to get granted policies to prevent outdated policies
