+++ apps/web/src/app/[lang]/(main)/(unirefund)/parties/individuals/table.tsx
@@ -4,7 +4,7 @@ import type { GetApiCrmServiceIndividualsResponse } from "@ayasofyazilim/saas/CR
 import TanstackTable from "@repo/ayasofyazilim-ui/molecules/tanstack-table";
 import { useParams, useRouter } from "next/navigation";
 import type { CRMServiceServiceResource } from "src/language-data/unirefund/CRMService";
-import useGrantedPolicies from "src/hooks/use-granted-policies";
+import { useGrantedPolicies } from "src/providers/granted-policies";
 import { tableData } from "./individuals-table-data";
 
 function IndividualsTable({
@@ -14,7 +14,7 @@ function IndividualsTable({
   response: GetApiCrmServiceIndividualsResponse;
   languageData: CRMServiceServiceResource;
 }) {
-  const grantedPolicies = useGrantedPolicies();
+  const { grantedPolicies } = useGrantedPolicies();
   const router = useRouter();
   const { lang } = useParams<{ lang: string }>();
   const columns = tableData.individuals.columns(
