+++ apps/web/src/app/[lang]/(main)/(unirefund)/parties/merchants/table.tsx
@@ -3,8 +3,8 @@
 import type { PagedResultDto_MerchantProfileDto } from "@ayasofyazilim/saas/CRMService";
 import TanstackTable from "@repo/ayasofyazilim-ui/molecules/tanstack-table";
 import { useParams, useRouter } from "next/navigation";
-import useGrantedPolicies from "src/hooks/use-granted-policies";
 import type { CRMServiceServiceResource } from "src/language-data/unirefund/CRMService";
+import { useGrantedPolicies } from "src/providers/granted-policies";
 import { tableData } from "./merchant-table-data";
 
 function MerchantsTable({
@@ -14,7 +14,7 @@ function MerchantsTable({
   response: PagedResultDto_MerchantProfileDto;
   languageData: CRMServiceServiceResource;
 }) {
-  const grantedPolicies = useGrantedPolicies();
+  const { grantedPolicies } = useGrantedPolicies();
   const router = useRouter();
   const { lang } = useParams<{ lang: string }>();
   const columns = tableData.merchants.columns(
