+++ apps/web/src/app/[lang]/(main)/(unirefund)/parties/refund-points/table.tsx
@@ -4,7 +4,7 @@ import type { PagedResultDto_RefundPointProfileDto } from "@ayasofyazilim/saas/C
 import TanstackTable from "@repo/ayasofyazilim-ui/molecules/tanstack-table";
 import { useParams, useRouter } from "next/navigation";
 import type { CRMServiceServiceResource } from "src/language-data/unirefund/CRMService";
-import useGrantedPolicies from "src/hooks/use-granted-policies";
+import { useGrantedPolicies } from "src/providers/granted-policies";
 import { tableData } from "./refund-points-table-data";
 
 function RefundPointsTable({
@@ -14,7 +14,7 @@ function RefundPointsTable({
   response: PagedResultDto_RefundPointProfileDto;
   languageData: CRMServiceServiceResource;
 }) {
-  const grantedPolicies = useGrantedPolicies();
+  const { grantedPolicies } = useGrantedPolicies();
   const router = useRouter();
   const { lang } = useParams<{ lang: string }>();
   const columns = tableData.refundPoints.columns(
