+++ apps/web/src/app/[lang]/(main)/(unirefund)/parties/travellers/table.tsx
@@ -4,7 +4,7 @@ import type { PagedResultDto_TravellerListProfileDto } from "@ayasofyazilim/saas
 import TanstackTable from "@repo/ayasofyazilim-ui/molecules/tanstack-table";
 import { useParams, useRouter } from "next/navigation";
 import type { TravellerServiceResource } from "src/language-data/unirefund/TravellerService";
-import useGrantedPolicies from "src/hooks/use-granted-policies";
+import { useGrantedPolicies } from "src/providers/granted-policies";
 import type { CountryDto } from "../../../../../../actions/unirefund/LocationService/types";
 import { tableData } from "./travellers-table-data";
 
@@ -17,7 +17,7 @@ function TravellersTable({
   languageData: TravellerServiceResource;
   countryList: CountryDto[];
 }) {
-  const grantedPolicies = useGrantedPolicies();
+  const { grantedPolicies } = useGrantedPolicies();
   const router = useRouter();
   const { lang } = useParams<{ lang: string }>();
   const columns = tableData.travellers.columns(
