+++ apps/web/src/app/[lang]/(main)/(unirefund)/refund/table.tsx
@@ -4,7 +4,7 @@ import type { PagedResultDto_RefundListItem } from "@ayasofyazilim/saas/RefundSe
 import TanstackTable from "@repo/ayasofyazilim-ui/molecules/tanstack-table";
 import { useRouter } from "next/navigation";
 import type { ContractServiceResource } from "src/language-data/unirefund/ContractService";
-import useGrantedPolicies from "src/hooks/use-granted-policies";
+import { useGrantedPolicies } from "src/providers/granted-policies";
 import { tableData } from "./table-data";
 
 function RefundTable({
@@ -16,7 +16,7 @@ function RefundTable({
   response: PagedResultDto_RefundListItem;
   languageData: ContractServiceResource;
 }) {
-  const grantedPolicies = useGrantedPolicies();
+  const { grantedPolicies } = useGrantedPolicies();
   const router = useRouter();
   const columns = tableData.refund.columns(locale);
   const table = tableData.refund.table(languageData, router, grantedPolicies);
